Add and chain ConfigureAwait(false) on all awaits to support blocking on async calls without deadlocking in .NET Framework applications
For general-purpose library code, adding ConfigureAwait(false) for every await allows .NET Framework applications that have a SynchronizationContext to call async code synchronously without deadlocking. The 'false' argument to ConfigureAwait allows the continuation to complete on any available thread in the pool and not just the original one. Since the context of the application is unknown to the library it can't use it anyway. To prevent future changes from reintroducing await's without a ConfigureAwait(false), the .editorconfig has been updated to treat these as errors and EnableNETAnalyzers added to the csproj. Tests are excluded from this rule as they are not expected to have a SynchronizationContext.

How does this affect calling applications:

* No impact on .NET Core apps. Async or synchronous calling is supported.
* No impact on .NET Framework apps that do not rely on a SynchronizationContext. Async or synchronous calling is supported.
* .NET Framework apps using SynchronizationContext (ASP.NET, WPF, Winforms) will no longer deadlock on synchronous calls as long as all await'ed async calls include ConfigureAwait(false).

Reference: https://devblogs.microsoft.com/dotnet/configureawait-faq/#when-should-i-use-configureawait(false)